requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain
langchain-google-genai
langchain-core
qdrant-client==1.7.3
numpy
sentence-transformers